requirements.txt
numpy==1.21.5
Pillow==8.4.0
python==3.10.0